Thread deleting Auth Policy

+++ app/Http/Controllers/ThreadsController.php
@@ -31,8 +31,7 @@ class ThreadsController extends Controller
     {
         $threads = $this->getThreads($channel, $filters);
 
-        if (\request()->wantsJson())
-        {
+        if (\request()->wantsJson()) {
             return $threads;
         }
 
@@ -121,7 +120,13 @@ class ThreadsController extends Controller
     public function destroy(Channel $channel, Thread $thread)
     {
 
+        $this->authorize('update', $thread);
+
         $thread->delete();
+
+        if (\request()->wantsJson()) {
+            return response([], 204);
+        }
 
         return redirect('/threads');
     }
